+++ pymarkdown/plugins/rule_md_005.py
@@ -4,6 +4,7 @@ are equivalent with each other.
 """
 from enum import Enum
 
+# from pymarkdown.parser_helper import ParserHelper
 from pymarkdown.plugin_details import PluginDetails
 from pymarkdown.rule_plugin import RulePlugin
 
@@ -31,6 +32,7 @@ class RuleMd005(RulePlugin):
         self.__ordered_list_starts = {}
         self.__ordered_tokens = {}
         self.__ordered_list_alignment = {}
+        # self.__debug = False
 
     def get_details(self):
         """
@@ -56,42 +58,18 @@ class RuleMd005(RulePlugin):
         self.__ordered_tokens = {}
         self.__ordered_list_alignment = {}
 
-    def __report_issue(self, context, token):
-        # print(f"token>>{ParserHelper.make_value_visible(token)}")
-        # print(f"self.__list_stack[]>>{ParserHelper.make_value_visible(self.__list_stack)}")
-        # show_item = str(self.__list_stack[-1]).replace(ParserHelper.newline_character, "\\n")
-        # print(f"self.__list_stack[-1]>>{show_item}")
-        if self.__list_stack[-1].is_unordered_list_start:
-            delta = self.__list_stack[-1].indent_level - 2
-        else:
-            delta = (
-                self.__list_stack[-1].indent_level
-                - 2
-                - len(self.__list_stack[-1].list_start_content)
-            )
-        current_list_indent = len(self.__list_stack[-1].extracted_whitespace) + delta
-        if self.__list_stack[0].is_unordered_list_start:
-            indent_adjust = self.__list_stack[0].column_number - 1
-        else:
-            indent_adjust = -1
-        token_indent = len(token.extracted_whitespace)
-        # print(f"token_indent>>{token_indent}")
-        token_indent -= indent_adjust
-        # print(f"token_indent>>{token_indent}")
-        if (
-            token_indent <= current_list_indent
-            and len(self.__list_stack) > 1
-            and self.__list_stack[-2].is_list_start
-        ):
-            # show_item = str(self.__list_stack[-2]).replace(ParserHelper.newline_character, "\\n")
-            # print(f"self.__list_stack[-2]>>{show_item}")
+    def __report_issue(self, context, token, expected_indent):
+        if expected_indent < 0:
+            list_level = len(self.__list_stack)
             expected_indent = (
-                len(self.__list_stack[-2].extracted_whitespace)
-                + self.__list_stack[-2].indent_level
+                self.__ordered_list_starts[list_level].indent_level
+                - 2
+                - len(token.list_start_content)
             )
-        else:
-            expected_indent = current_list_indent
-        extra_data = f"Expected: {expected_indent}; Actual: {token_indent}"
+
+        extra_data = f"Expected: {expected_indent}; Actual: {token.column_number - 1}"
+        # if self.__debug:
+        #     print(f"ERROR>>{extra_data}")
         self.report_next_token_error(context, token, extra_data)
 
     def __handle_ordered_list_item(self, context, token):
@@ -114,11 +92,17 @@ class RuleMd005(RulePlugin):
                     == self.__ordered_list_starts[list_level].indent_level
                 )
             else:
-                self.__report_issue(context, token)
+                # if self.__debug:
+                #     print("ri1")
+                self.__report_issue(context, token, -1)
         elif (
             self.__ordered_list_starts[list_level].column_number != token.column_number
         ):
-            self.__report_issue(context, token)
+            # if self.__debug:
+            #     print("ri2")
+            self.__report_issue(
+                context, token, self.__ordered_list_starts[list_level].column_number - 1
+            )
 
     def __compute_ordered_list_alignment(self):
 
@@ -156,7 +140,11 @@ class RuleMd005(RulePlugin):
         if list_level not in self.__unordered_list_indents:
             self.__unordered_list_indents[list_level] = token.indent_level
         if self.__unordered_list_indents[list_level] != token.indent_level:
-            self.__report_issue(context, token)
+            # if self.__debug:
+            #     print("ri3")
+            self.__report_issue(
+                context, token, self.__unordered_list_indents[list_level] - 2
+            )
 
     def __handle_ordered_list_start(self, token):
         self.__list_stack.append(token)
@@ -173,7 +161,15 @@ class RuleMd005(RulePlugin):
                 self.__unordered_list_indents[len(self.__list_stack)]
                 != token.indent_level
             ):
-                self.__report_issue(context, token)
+                # if self.__debug:
+                #     print("ri4")
+                #     print("token.indent_level=" + str(token.indent_level))
+                #     print("un_li=" + ParserHelper.make_value_visible(self.__unordered_list_indents))
+                self.__report_issue(
+                    context,
+                    token,
+                    self.__unordered_list_indents[len(self.__list_stack)] - 2,
+                )
         else:
             self.__ordered_tokens[len(self.__list_stack)].append(token)