+++ test/rules/test_md005.py
@@ -278,7 +278,7 @@ def test_md005_bad_ordered_list_single_level_x():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_single_level.md:2:2: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 0; Actual: 2] (list-indent)"
+        + "[Expected: 0; Actual: 1] (list-indent)"
     )
     expected_error = ""
 
@@ -342,7 +342,7 @@ def test_md005_bad_ordered_list_single_level_widths():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_single_level_widths.md:2:2: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 0; Actual: 2] (list-indent)"
+        + "[Expected: 0; Actual: 1] (list-indent)"
     )
     expected_error = ""
 
@@ -405,7 +405,7 @@ def test_md005_bad_ordered_list_single_level_widths_right():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_single_level_widths_right.md:2:1: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 6; Actual: 1] (list-indent)"
+        + "[Expected: 3; Actual: 0] (list-indent)"
     )
     expected_error = ""
 
@@ -583,7 +583,7 @@ def test_md005_bad_ordered_list_double_level_weirdx():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_double_level_weird.md:3:5: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 3; Actual: 5] (list-indent)"
+        + "[Expected: 3; Actual: 4] (list-indent)"
     )
     expected_error = ""
 
@@ -616,7 +616,8 @@ def test_md005_bad_ordered_list_double_level_weirder():
     expected_return_code = 1
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_double_level_weirder.md:3:3: "
-        + "MD005: Inconsistent indentation for list items at the same level [Expected: 0; Actual: 3] (list-indent)"
+        + "MD005: Inconsistent indentation for list items at the same level "
+        + "[Expected: 0; Actual: 2] (list-indent)"
     )
     expected_error = ""
 
@@ -676,7 +677,7 @@ def test_md005_good_unordered_list_double_level_in_block_quote_first():
     expected_output = (
         "test/resources/rules/md005/bad_unordered_list_double_level_in_block_quote_first.md:4:4: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 2; Actual: -1] (list-indent)"
+        + "[Expected: 2; Actual: 3] (list-indent)"
     )
     expected_error = ""
 
@@ -709,10 +710,10 @@ def test_md005_good_unordered_list_double_level_in_block_quote_second():
     expected_output = (
         "test/resources/rules/md005/bad_unordered_list_double_level_in_block_quote_second.md:5:6: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 4; Actual: 1] (list-indent)\n"
+        + "[Expected: 4; Actual: 5] (list-indent)\n"
         + "test/resources/rules/md005/bad_unordered_list_double_level_in_block_quote_second.md:6:6: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 4; Actual: 1] (list-indent)"
+        + "[Expected: 4; Actual: 5] (list-indent)"
     )
     expected_error = ""
 
@@ -746,10 +747,10 @@ def test_md005_bad_ordered_list_double_level_left():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_double_level_left.md:5:5: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 3; Actual: 5] (list-indent)\n"
+        + "[Expected: 3; Actual: 4] (list-indent)\n"
         + "test/resources/rules/md005/bad_ordered_list_double_level_left.md:6:5: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 3; Actual: 5] (list-indent)"
+        + "[Expected: 3; Actual: 4] (list-indent)"
     )
     expected_error = ""
 
@@ -763,7 +764,7 @@ def test_md005_bad_ordered_list_double_level_left():
 
 
 @pytest.mark.rules
-def test_md005_bad_ordered_list_double_level_right():
+def test_md005_bad_ordered_list_double_level_right_x():
     """
     Test to make sure this rule does trigger with a document that
     has two level 1 ordered lists with right alignment and bad indentation
@@ -783,10 +784,10 @@ def test_md005_bad_ordered_list_double_level_right():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_double_level_right.md:5:6: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 3; Actual: 6] (list-indent)\n"
+        + "[Expected: 4; Actual: 5] (list-indent)\n"
         + "test/resources/rules/md005/bad_ordered_list_double_level_right.md:6:5: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 3; Actual: 5] (list-indent)"
+        + "[Expected: 3; Actual: 4] (list-indent)"
     )
     expected_error = ""
 
@@ -820,7 +821,7 @@ def test_md005_bad_ordered_list_double_level_left_then_right():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_double_level_left_then_right.md:5:5: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 3; Actual: 5] (list-indent)"
+        + "[Expected: 3; Actual: 4] (list-indent)"
     )
     expected_error = ""
 
@@ -854,7 +855,7 @@ def test_md005_bad_ordered_list_double_level_right_then_left():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_double_level_right_then_left.md:5:4: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 3; Actual: 4] (list-indent)"
+        + "[Expected: 4; Actual: 3] (list-indent)"
     )
     expected_error = ""
 
@@ -888,7 +889,7 @@ def test_md005_bad_ordered_list_single_level_left_then_right():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_single_level_left_then_right.md:3:2: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 0; Actual: 2] (list-indent)"
+        + "[Expected: 0; Actual: 1] (list-indent)"
     )
     expected_error = ""
 
@@ -914,6 +915,7 @@ def test_md005_bad_ordered_list_single_level_right_then_left():
     supplied_arguments = [
         "--disable-rules",
         "md029",
+        "--stack-trace",
         "scan",
         "test/resources/rules/md005/bad_ordered_list_single_level_right_then_left.md",
     ]
@@ -922,7 +924,7 @@ def test_md005_bad_ordered_list_single_level_right_then_left():
     expected_output = (
         "test/resources/rules/md005/bad_ordered_list_single_level_right_then_left.md:3:1: "
         + "MD005: Inconsistent indentation for list items at the same level "
-        + "[Expected: 2; Actual: 1] (list-indent)"
+        + "[Expected: 1; Actual: 0] (list-indent)"
     )
     expected_error = ""