+++ test/test_markdown_extra.py
@@ -2568,8 +2568,7 @@ def test_extra_024a():
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
-def test_extra_025x():
+def test_extra_025xx():
     """
     TBD
     """
@@ -2582,34 +2581,21 @@ def test_extra_025x():
 > + that
 """
     expected_tokens = [
-        "[block-quote(1,1)::> \n> \n> \n> \n> \n> ]",
-        "[ulist(1,3):*::4:]",
-        "[para(1,5):]",
-        "[text(1,5):Item 1:]",
+        "[block-quote(1,1)::> \n> \n> ]",
+        "[ulist(1,3):+::4::  \n\n\n]",
+        "[para(1,5):\n]",
+        "[text(1,5):list\nthis::\n]",
         "[end-para:::True]",
-        "[ulist(2,5):*::6:  ]",
-        "[para(2,7):]",
-        "[text(2,7):Item 1a:]",
+        "[block-quote(3,5)::> \n>   > \n> ]",
+        "[para(3,8): \n]",
+        "[text(3,8):good\nitem::\n]",
         "[end-para:::True]",
-        "[li(3,5):6:  :]",
-        "[para(3,7):]",
-        "[text(3,7):Item 1b:]",
+        "[end-block-quote:::True]",
+        "[li(5,3):4:  :]",
+        "[para(5,5):]",
+        "[text(5,5):that:]",
         "[end-para:::True]",
-        "[end-ulist:::True]",
-        "[li(4,4):5: :]",
-        "[para(4,6):]",
-        "[text(4,6):Item 2:]",
-        "[end-para:::True]",
-        "[ulist(5,6):*::7:   :]",
-        "[para(5,8):]",
-        "[text(5,8):Item 2a:]",
-        "[end-para:::True]",
-        "[li(6,6):7:   :]",
-        "[para(6,8):]",
-        "[text(6,8):Item 2b:]",
-        "[end-para:::True]",
-        "[BLANK(7,1):]",
-        "[end-ulist:::True]",
+        "[BLANK(6,1):]",
         "[end-ulist:::True]",
         "[end-block-quote:::True]",
     ]
@@ -2631,8 +2617,52 @@ item</p>
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
-def test_extra_025a():
+def test_extra_025xa():
+    """
+    TBD
+    """
+
+    # Arrange
+    source_markdown = """> + list
+>   this
+>   >  good
+>   > item
+>   that
+"""
+    expected_tokens = [
+        "[block-quote(1,1)::> \n> \n> ]",
+        "[ulist(1,3):+::4::  \n\n\n\n]",
+        "[para(1,5):\n]",
+        "[text(1,5):list\nthis::\n]",
+        "[end-para:::True]",
+        "[block-quote(3,5)::> \n>   > \n> \n]",
+        "[para(3,8): \n\n  ]",
+        "[text(3,8):good\nitem\nthat::\n\n]",
+        "[end-para:::True]",
+        "[end-block-quote:::True]",
+        "[BLANK(6,1):]",
+        "[end-ulist:::True]",
+        "[end-block-quote:::True]",
+    ]
+    expected_gfm = """<blockquote>
+<ul>
+<li>list
+this
+<blockquote>
+<p>good
+item
+that</p>
+</blockquote>
+</li>
+</ul>
+</blockquote>"""
+
+    # Act & Assert
+    act_and_assert(source_markdown, expected_gfm, expected_tokens)
+
+
+@pytest.mark.gfm
+def test_extra_025ax():
     """
     TBD
     """
@@ -2645,34 +2675,21 @@ def test_extra_025a():
 > + that
 """
     expected_tokens = [
-        "[block-quote(1,1)::> \n> \n> \n> \n> \n> ]",
-        "[ulist(1,3):*::4:]",
-        "[para(1,5):]",
-        "[text(1,5):Item 1:]",
+        "[block-quote(1,1)::> \n> \n> ]",
+        "[ulist(1,3):+::4::  \n\n\n]",
+        "[para(1,5):\n]",
+        "[text(1,5):list\nthis::\n]",
         "[end-para:::True]",
-        "[ulist(2,5):*::6:  ]",
-        "[para(2,7):]",
-        "[text(2,7):Item 1a:]",
+        "[block-quote(3,5)::> \n>   > \n> ]",
+        "[para(3,7):\n]",
+        "[text(3,7):good\nitem::\n]",
         "[end-para:::True]",
-        "[li(3,5):6:  :]",
-        "[para(3,7):]",
-        "[text(3,7):Item 1b:]",
+        "[end-block-quote:::True]",
+        "[li(5,3):4:  :]",
+        "[para(5,5):]",
+        "[text(5,5):that:]",
         "[end-para:::True]",
-        "[end-ulist:::True]",
-        "[li(4,4):5: :]",
-        "[para(4,6):]",
-        "[text(4,6):Item 2:]",
-        "[end-para:::True]",
-        "[ulist(5,6):*::7:   :]",
-        "[para(5,8):]",
-        "[text(5,8):Item 2a:]",
-        "[end-para:::True]",
-        "[li(6,6):7:   :]",
-        "[para(6,8):]",
-        "[text(6,8):Item 2b:]",
-        "[end-para:::True]",
-        "[BLANK(7,1):]",
-        "[end-ulist:::True]",
+        "[BLANK(6,1):]",
         "[end-ulist:::True]",
         "[end-block-quote:::True]",
     ]
@@ -2694,8 +2711,52 @@ item</p>
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
-def test_extra_025b():
+def test_extra_025aa():
+    """
+    TBD
+    """
+
+    # Arrange
+    source_markdown = """> + list
+>   this
+>   > good
+>   > item
+>   that
+"""
+    expected_tokens = [
+        "[block-quote(1,1)::> \n> \n> ]",
+        "[ulist(1,3):+::4::  \n\n\n\n]",
+        "[para(1,5):\n]",
+        "[text(1,5):list\nthis::\n]",
+        "[end-para:::True]",
+        "[block-quote(3,5)::> \n>   > \n> \n]",
+        "[para(3,7):\n\n  ]",
+        "[text(3,7):good\nitem\nthat::\n\n]",
+        "[end-para:::True]",
+        "[end-block-quote:::True]",
+        "[BLANK(6,1):]",
+        "[end-ulist:::True]",
+        "[end-block-quote:::True]",
+    ]
+    expected_gfm = """<blockquote>
+<ul>
+<li>list
+this
+<blockquote>
+<p>good
+item
+that</p>
+</blockquote>
+</li>
+</ul>
+</blockquote>"""
+
+    # Act & Assert
+    act_and_assert(source_markdown, expected_gfm, expected_tokens)
+
+
+@pytest.mark.gfm
+def test_extra_025bx():
     """
     TBD
     """
@@ -2742,8 +2803,50 @@ this
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
-def test_extra_025c():
+def test_extra_025ba():
+    """
+    TBD
+    """
+
+    # Arrange
+    source_markdown = """> + list
+>   this
+>   > item
+>   that
+"""
+    expected_tokens = [
+        "[block-quote(1,1)::> \n> \n> ]",
+        "[ulist(1,3):+::4::  \n\n\n]",
+        "[para(1,5):\n]",
+        "[text(1,5):list\nthis::\n]",
+        "[end-para:::True]",
+        "[block-quote(3,5)::> \n> \n]",
+        "[para(3,7):\n  ]",
+        "[text(3,7):item\nthat::\n]",
+        "[end-para:::True]",
+        "[end-block-quote:::True]",
+        "[BLANK(5,1):]",
+        "[end-ulist:::True]",
+        "[end-block-quote:::True]",
+    ]
+    expected_gfm = """<blockquote>
+<ul>
+<li>list
+this
+<blockquote>
+<p>item
+that</p>
+</blockquote>
+</li>
+</ul>
+</blockquote>"""
+
+    # Act & Assert
+    act_and_assert(source_markdown, expected_gfm, expected_tokens)
+
+
+@pytest.mark.gfm
+def test_extra_025cx():
     """
     TBD
     """
@@ -2755,17 +2858,17 @@ def test_extra_025c():
 > + that
 """
     expected_tokens = [
-        "[block-quote(1,1)::> \n> \n> ]",
-        "[ulist(1,3):+::4::  \n\n]",
-        "[para(1,5):\n]",
-        "[text(1,5):list\nthis::\n]",
+        "[block-quote(1,1)::> \n> ]",
+        "[ulist(1,3):+::4::\n\n]",
+        "[para(1,5):]",
+        "[text(1,5):list:]",
         "[end-para:::True]",
-        "[block-quote(3,5)::> \n> ]",
-        "[para(3,7):]",
-        "[text(3,7):item:]",
+        "[block-quote(2,5)::> \n>   > \n> ]",
+        "[para(2,7):\n]",
+        "[text(2,7):good\nitem::\n]",
         "[end-para:::True]",
         "[end-block-quote:::True]",
-        "[li(4,3):4::]",
+        "[li(4,3):4:  :]",
         "[para(4,5):]",
         "[text(4,5):that:]",
         "[end-para:::True]",
@@ -2776,12 +2879,55 @@ def test_extra_025c():
     expected_gfm = """<blockquote>
 <ul>
 <li>list
-this
 <blockquote>
-<p>item</p>
+<p>good
+item</p>
 </blockquote>
 </li>
 <li>that</li>
+</ul>
+</blockquote>"""
+
+    # Act & Assert
+    act_and_assert(source_markdown, expected_gfm, expected_tokens, show_debug=False)
+
+
+@pytest.mark.gfm
+def test_extra_025ca():
+    """
+    TBD
+    """
+
+    # Arrange
+    source_markdown = """> + list
+>   > good
+>   > item
+>   that
+"""
+    expected_tokens = [
+        "[block-quote(1,1)::> \n> ]",
+        "[ulist(1,3):+::4::\n\n\n]",
+        "[para(1,5):]",
+        "[text(1,5):list:]",
+        "[end-para:::True]",
+        "[block-quote(2,5)::> \n>   > \n> \n]",
+        "[para(2,7):\n\n  ]",
+        "[text(2,7):good\nitem\nthat::\n\n]",
+        "[end-para:::True]",
+        "[end-block-quote:::True]",
+        "[BLANK(5,1):]",
+        "[end-ulist:::True]",
+        "[end-block-quote:::True]",
+    ]
+    expected_gfm = """<blockquote>
+<ul>
+<li>list
+<blockquote>
+<p>good
+item
+that</p>
+</blockquote>
+</li>
 </ul>
 </blockquote>"""