+++ changelog.md
@@ -8,11 +8,17 @@
 
 ### Changed
 
-- None
+- [Changed - Issue 154](https://github.com/jackdewinter/pymarkdown/issues/154)
+  - rule md003 - added configuration `allow-setext-update`
 
 ### Fixed
 
-- None
+- [Fixed - Issue 95](https://github.com/jackdewinter/pymarkdown/issues/95)
+  - parser: with certain cases involving a new list item, can start a container in the middle of line
+- [Fixed - Issue 161](https://github.com/jackdewinter/pymarkdown/issues/161)
+  - rule md005: was not returning proper values for actual and expected
+- [Fixed - Issue 189](https://github.com/jackdewinter/pymarkdown/issues/189)
+  - rule md027: addressed index out of bounds error keeping track of blank lines after a block quote was started in a previous section
 
 ## Version 0.9.5 - Date: 2022-02-07